Fix CI for macOS and iOS
